requirements.txt
# Production requirements for Stray Mapper - Updated to Django 4.2 LTS
Django==4.2.16
Pillow==11.0.0
celery==5.5.3
django-celery-beat==2.8.0
django-el-pagination==4.1.2
django-extensions==3.2.3
django-imagekit==5.0.0
geopy==2.4.1
gunicorn==23.0.0
django-storages==1.14.4
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
